Added methods to handle tasks to App
Moved Task and Priority to task.rs
