+++ src/app.rs
@@ -1,5 +1,6 @@
 use ratatui::widgets::ListState;
-use crate::task_manager::{Priority, Task};
+use crate::task::{Priority, Task};
+use crate::task_manager::TasksService;
 
 pub struct TaskList {
     pub items: Vec<Task>,
@@ -25,6 +26,7 @@ pub struct App {
     pub task_list: TaskList,
     pub input: String,
     pub input_mode: InputMode,
+    pub tasks_service: TasksService
 }
 
 impl App {
@@ -33,6 +35,7 @@ impl App {
             task_list: TaskList::new(),
             input: String::new(),
             input_mode: InputMode::Normal,
+            tasks_service: TasksService::new(&"tasks.db".to_string()),
         }
     }
 
@@ -53,5 +56,70 @@ impl App {
           },
         ];
         app
+    }
+
+    /// Toggle completed for selected task status
+    pub fn toggle_item_status(&mut self) {
+        if let Some(index) = self.task_list.state.selected(){
+            let item = &mut self.task_list.items[index];
+            item.completed = match item.completed {
+                true => false,
+                false => true,
+            }
+        };
+    }
+
+    /// Cycle through priorities
+    pub fn toggle_item_priority(&mut self) {
+        if let Some(index) = self.task_list.state.selected() {
+            let item = &mut self.task_list.items[index];
+            item.priority = item.priority.next();
+        }
+    }
+
+    /// Copy task description into App.input and set the ui to InputMode::EditingExisting
+    pub fn start_editing_exisiting(&mut self) {
+        if let Some(index) = self.task_list.state.selected() {
+            self.input = self.task_list.items[index].description.clone();
+            self.input_mode = InputMode::EditingExisting;
+        }
+    }
+
+    /// Copy new description into the selected item and set the ui to InputMode::Normal
+    pub fn finish_editing_existing(&mut self) {
+        if let Some(index) = self.task_list.state.selected() {
+            self.task_list.items[index].description = self.input.drain(..).collect();
+        }
+        self.input_mode = InputMode::Normal;
+    }
+
+    pub fn sort_by_priority(&mut self) {
+        self.task_list.items.sort_by(|a, b| b.priority.cmp(&a.priority))
+    }
+
+    pub fn delete_item(&mut self) {
+        if let Some(index) = self.task_list.state.selected() {
+            self.task_list.items.remove(index);
+        }
+    }
+
+    pub fn select_none(&mut self) {
+        self.task_list.state.select(None);
+    }
+
+    pub fn select_next(&mut self) {
+        self.task_list.state.select_next();
+    }
+
+    pub fn select_previous(&mut self) {
+        self.task_list.state.select_previous();
+    }
+
+    pub fn select_first(&mut self) {
+        self.task_list.state.select_first();
+    }
+
+    pub fn select_last(&mut self) {
+        self.task_list.state.select_last();
     }
 }
