+++ src/lib.rs
@@ -2,5 +2,6 @@ pub mod task_db;
 pub mod task_manager;
 pub mod app;
 pub mod ui;
+pub mod task;
